requirements.txt
django~=4.1
pymysql~=1.1.0
djangorestframework~=3.14.0
python-jenkins~=1.8.0
celery~=5.3.1